@@ -2,7 +2,7 @@
 import DequeModule
 
 @available(macOS 13.0, iOS 16.0, tvOS 16.0, watchOS 9.0, *)
-final class MockConnectionFactory<Clock: _Concurrency.Clock> where Clock.Duration == Duration {
+final class MockConnectionFactory<Clock: _Concurrency.Clock>: Sendable where Clock.Duration == Duration {
     typealias ConnectionIDGenerator = _ConnectionPoolModule.ConnectionIDGenerator
     typealias Request = ConnectionRequest<MockConnection>
     typealias KeepAliveBehavior = MockPingPongBehavior
