@@ -25,12 +25,49 @@ final class PostgresClientTests: XCTestCase {
                 await client.run()
             }
 
-            for i in 0..<10000 {
+            let iterations = 1000
+
+            for _ in 0..<iterations {
                 taskGroup.addTask {
                     try await client.withConnection() { connection in
                         _ = try await connection.query("SELECT 1", logger: logger)
                     }
-                    print("done: \(i)")
+                }
+            }
+
+            for _ in 0..<iterations {
+                _ = await taskGroup.nextResult()!
+            }
+
+            taskGroup.cancelAll()
+        }
+    }
+
+    func testQueryDirectly() async throws {
+        var mlogger = Logger(label: "test")
+        mlogger.logLevel = .debug
+        let logger = mlogger
+        let eventLoopGroup = MultiThreadedEventLoopGroup(numberOfThreads: 8)
+        self.addTeardownBlock {
+            try await eventLoopGroup.shutdownGracefully()
+        }
+
+        let clientConfig = PostgresClient.Configuration.makeTestConfiguration()
+        let client = PostgresClient(configuration: clientConfig, eventLoopGroup: eventLoopGroup, backgroundLogger: logger)
+
+        await withThrowingTaskGroup(of: Void.self) { taskGroup in
+            taskGroup.addTask {
+                await client.run()
+            }
+
+            for i in 0..<10000 {
+                taskGroup.addTask {
+                    do {
+                        try await client.query("SELECT 1", logger: logger)
+                        logger.info("Success", metadata: ["run": "\(i)"])
+                    } catch {
+                        XCTFail("Unexpected error: \(error)")
+                    }
                 }
             }
 
@@ -39,6 +76,81 @@ final class PostgresClientTests: XCTestCase {
             }
 
             taskGroup.cancelAll()
+        }
+    }
+
+    func testQueryTable() async throws {
+        let tableName = "test_client_prepared_statement"
+
+        var mlogger = Logger(label: "test")
+        mlogger.logLevel = .debug
+        let logger = mlogger
+        let eventLoopGroup = MultiThreadedEventLoopGroup(numberOfThreads: 8)
+        self.addTeardownBlock {
+            try await eventLoopGroup.shutdownGracefully()
+        }
+
+        let clientConfig = PostgresClient.Configuration.makeTestConfiguration()
+        let client = PostgresClient(configuration: clientConfig, eventLoopGroup: eventLoopGroup, backgroundLogger: logger)
+        do {
+            try await withThrowingTaskGroup(of: Void.self) { taskGroup in
+                taskGroup.addTask {
+                    await client.run()
+                }
+
+                try await client.query(
+                    """
+                    CREATE TABLE IF NOT EXISTS "\(unescaped: tableName)" (
+                        id SERIAL PRIMARY KEY,
+                        uuid UUID NOT NULL
+                    );
+                    """,
+                    logger: logger
+                )
+
+                for _ in 0..<1000 {
+                    try await client.query(
+                        """
+                        INSERT INTO "\(unescaped: tableName)" (uuid) VALUES (\(UUID()));
+                        """,
+                        logger: logger
+                    )
+                }
+
+                let rows = try await client.query(#"SELECT id, uuid FROM "\#(unescaped: tableName)";"#, logger: logger).decode((Int, UUID).self)
+                for try await (id, uuid) in rows {
+                    logger.info("id: \(id), uuid: \(uuid.uuidString)")
+                }
+
+                struct Example: PostgresPreparedStatement {
+                    static let sql = "SELECT id, uuid FROM test_client_prepared_statement WHERE id < $1"
+                    typealias Row = (Int, UUID)
+                    var id: Int
+                    func makeBindings() -> PostgresBindings {
+                        var bindings = PostgresBindings()
+                        bindings.append(self.id)
+                        return bindings
+                    }
+                    func decodeRow(_ row: PostgresNIO.PostgresRow) throws -> Row {
+                        try row.decode(Row.self)
+                    }
+                }
+
+                for try await (id, uuid) in try await client.execute(Example(id: 200), logger: logger) {
+                    logger.info("id: \(id), uuid: \(uuid.uuidString)")
+                }
+
+                try await client.query(
+                    """
+                    DROP TABLE "\(unescaped: tableName)";
+                    """,
+                    logger: logger
+                )
+
+                taskGroup.cancelAll()
+            }
+        } catch {
+            XCTFail("Unexpected error: \(String(reflecting: error))")
         }
     }
 }
