@@ -38,6 +38,48 @@ class PostgresConnectionTests: XCTestCase {
         }
     }
 
+    func testOptionsAreSentOnTheWire() async throws {
+        let eventLoop = NIOAsyncTestingEventLoop()
+        let channel = await NIOAsyncTestingChannel(handlers: [
+            ReverseByteToMessageHandler(PSQLFrontendMessageDecoder()),
+            ReverseMessageToByteHandler(PSQLBackendMessageEncoder()),
+        ], loop: eventLoop)
+        try await channel.connect(to: .makeAddressResolvingHost("localhost", port: 5432))
+
+        let configuration = {
+            var config = PostgresConnection.Configuration(
+                establishedChannel: channel,
+                username: "username",
+                password: "postgres",
+                database: "database"
+            )
+            config.options.additionalStartupParameters = [
+                ("DateStyle", "ISO, MDY"),
+                ("application_name", "postgres-nio-test"),
+                ("server_encoding", "UTF8"),
+                ("integer_datetimes", "on"),
+                ("client_encoding", "UTF8"),
+                ("TimeZone", "Etc/UTC"),
+                ("is_superuser", "on"),
+                ("server_version", "13.1 (Debian 13.1-1.pgdg100+1)"),
+                ("session_authorization", "postgres"),
+                ("IntervalStyle", "postgres"),
+                ("standard_conforming_strings", "on")
+            ]
+            return config
+        }()
+
+        async let connectionPromise = PostgresConnection.connect(on: eventLoop, configuration: configuration, id: 1, logger: .psqlTest)
+        let message = try await channel.waitForOutboundWrite(as: PostgresFrontendMessage.self)
+        XCTAssertEqual(message, .startup(.versionThree(parameters: .init(user: "username", database: "database", options: configuration.options.additionalStartupParameters, replication: .false))))
+        try await channel.writeInbound(PostgresBackendMessage.authentication(.ok))
+        try await channel.writeInbound(PostgresBackendMessage.backendKeyData(.init(processID: 1234, secretKey: 5678)))
+        try await channel.writeInbound(PostgresBackendMessage.readyForQuery(.idle))
+
+        let connection = try await connectionPromise
+        try await connection.close()
+    }
+
     func testSimpleListen() async throws {
         let (connection, channel) = try await self.makeTestConnectionWithAsyncTestingChannel()
 
@@ -51,7 +93,7 @@ class PostgresConnectionTests: XCTestCase {
             }
 
             let listenMessage = try await channel.waitForUnpreparedRequest()
-            XCTAssertEqual(listenMessage.parse.query, "LISTEN foo;")
+            XCTAssertEqual(listenMessage.parse.query, #"LISTEN "foo";"#)
 
             try await channel.writeInbound(PostgresBackendMessage.parseComplete)
             try await channel.writeInbound(PostgresBackendMessage.parameterDescription(.init(dataTypes: [])))
@@ -63,7 +105,7 @@ class PostgresConnectionTests: XCTestCase {
             try await channel.writeInbound(PostgresBackendMessage.notification(.init(backendPID: 12, channel: "foo", payload: "wooohooo")))
 
             let unlistenMessage = try await channel.waitForUnpreparedRequest()
-            XCTAssertEqual(unlistenMessage.parse.query, "UNLISTEN foo;")
+            XCTAssertEqual(unlistenMessage.parse.query, #"UNLISTEN "foo";"#)
 
             try await channel.writeInbound(PostgresBackendMessage.parseComplete)
             try await channel.writeInbound(PostgresBackendMessage.parameterDescription(.init(dataTypes: [])))
@@ -111,7 +153,7 @@ class PostgresConnectionTests: XCTestCase {
             }
 
             let listenMessage = try await channel.waitForUnpreparedRequest()
-            XCTAssertEqual(listenMessage.parse.query, "LISTEN foo;")
+            XCTAssertEqual(listenMessage.parse.query, #"LISTEN "foo";"#)
 
             try await channel.writeInbound(PostgresBackendMessage.parseComplete)
             try await channel.writeInbound(PostgresBackendMessage.parameterDescription(.init(dataTypes: [])))
@@ -124,7 +166,7 @@ class PostgresConnectionTests: XCTestCase {
             try await channel.writeInbound(PostgresBackendMessage.notification(.init(backendPID: 12, channel: "foo", payload: "wooohooo2")))
 
             let unlistenMessage = try await channel.waitForUnpreparedRequest()
-            XCTAssertEqual(unlistenMessage.parse.query, "UNLISTEN foo;")
+            XCTAssertEqual(unlistenMessage.parse.query, #"UNLISTEN "foo";"#)
 
             try await channel.writeInbound(PostgresBackendMessage.parseComplete)
             try await channel.writeInbound(PostgresBackendMessage.parameterDescription(.init(dataTypes: [])))
@@ -145,7 +187,7 @@ class PostgresConnectionTests: XCTestCase {
     func testSimpleListenConnectionDrops() async throws {
         let (connection, channel) = try await self.makeTestConnectionWithAsyncTestingChannel()
 
-        try await withThrowingTaskGroup(of: Void.self) { taskGroup in
+        try await withThrowingTaskGroup(of: Void.self) { [logger] taskGroup in
             taskGroup.addTask {
                 let events = try await connection.listen("foo")
                 var iterator = events.makeAsyncIterator()
@@ -155,12 +197,12 @@ class PostgresConnectionTests: XCTestCase {
                     _ = try await iterator.next()
                     XCTFail("Did not expect to not throw")
                 } catch {
-                    print(error)
+                    logger.error("error", metadata: ["error": "\(error)"])
                 }
             }
 
             let listenMessage = try await channel.waitForUnpreparedRequest()
-            XCTAssertEqual(listenMessage.parse.query, "LISTEN foo;")
+            XCTAssertEqual(listenMessage.parse.query, #"LISTEN "foo";"#)
 
             try await channel.writeInbound(PostgresBackendMessage.parseComplete)
             try await channel.writeInbound(PostgresBackendMessage.parameterDescription(.init(dataTypes: [])))
@@ -184,10 +226,10 @@ class PostgresConnectionTests: XCTestCase {
 
     func testCloseGracefullyClosesWhenInternalQueueIsEmpty() async throws {
         let (connection, channel) = try await self.makeTestConnectionWithAsyncTestingChannel()
-        try await withThrowingTaskGroup(of: Void.self) { taskGroup async throws -> () in
+        try await withThrowingTaskGroup(of: Void.self) { [logger] taskGroup async throws -> () in
             for _ in 1...2 {
                 taskGroup.addTask {
-                    let rows = try await connection.query("SELECT 1;", logger: self.logger)
+                    let rows = try await connection.query("SELECT 1;", logger: logger)
                     var iterator = rows.decode(Int.self).makeAsyncIterator()
                     let first = try await iterator.next()
                     XCTAssertEqual(first, 1)
@@ -244,10 +286,10 @@ class PostgresConnectionTests: XCTestCase {
     func testCloseClosesImmediatly() async throws {
         let (connection, channel) = try await self.makeTestConnectionWithAsyncTestingChannel()
 
-        try await withThrowingTaskGroup(of: Void.self) { taskGroup async throws -> () in
+        try await withThrowingTaskGroup(of: Void.self) { [logger] taskGroup async throws -> () in
             for _ in 1...2 {
                 taskGroup.addTask {
-                    try await connection.query("SELECT 1;", logger: self.logger)
+                    try await connection.query("SELECT 1;", logger: logger)
                 }
             }
 
@@ -277,8 +319,9 @@ class PostgresConnectionTests: XCTestCase {
 
     func testIfServerJustClosesTheErrorReflectsThat() async throws {
         let (connection, channel) = try await self.makeTestConnectionWithAsyncTestingChannel()
+        let logger = self.logger
 
-        async let response = try await connection.query("SELECT 1;", logger: self.logger)
+        async let response = try await connection.query("SELECT 1;", logger: logger)
 
         let listenMessage = try await channel.waitForUnpreparedRequest()
         XCTAssertEqual(listenMessage.parse.query, "SELECT 1;")
@@ -337,7 +380,7 @@ class PostgresConnectionTests: XCTestCase {
 
             let prepareRequest = try await channel.waitForPrepareRequest()
             XCTAssertEqual(prepareRequest.parse.query, "SELECT datname FROM pg_stat_activity WHERE state = $1")
-            XCTAssertEqual(prepareRequest.parse.parameters.count, 0)
+            XCTAssertEqual(prepareRequest.parse.parameters.first, .text)
             guard case .preparedStatement(let name) = prepareRequest.describe else {
                 fatalError("Describe should contain a prepared statement")
             }
@@ -374,6 +417,16 @@ class PostgresConnectionTests: XCTestCase {
         }
     }
 
+    func testWeDontCrashOnUnexpectedChannelEvents() async throws {
+        let (connection, channel) = try await self.makeTestConnectionWithAsyncTestingChannel()
+
+        enum MyEvent {
+            case pleaseDontCrash
+        }
+        channel.pipeline.fireUserInboundEventTriggered(MyEvent.pleaseDontCrash)
+        try await connection.close()
+    }
+
     func testSerialExecutionOfSamePreparedStatement() async throws {
         let (connection, channel) = try await self.makeTestConnectionWithAsyncTestingChannel()
 
@@ -393,7 +446,7 @@ class PostgresConnectionTests: XCTestCase {
 
             let prepareRequest = try await channel.waitForPrepareRequest()
             XCTAssertEqual(prepareRequest.parse.query, "SELECT datname FROM pg_stat_activity WHERE state = $1")
-            XCTAssertEqual(prepareRequest.parse.parameters.count, 0)
+            XCTAssertEqual(prepareRequest.parse.parameters.first, .text)
             guard case .preparedStatement(let name) = prepareRequest.describe else {
                 fatalError("Describe should contain a prepared statement")
             }
@@ -487,7 +540,7 @@ class PostgresConnectionTests: XCTestCase {
             // The channel deduplicates prepare requests, we're going to see only one of them
             let prepareRequest = try await channel.waitForPrepareRequest()
             XCTAssertEqual(prepareRequest.parse.query, "SELECT datname FROM pg_stat_activity WHERE state = $1")
-            XCTAssertEqual(prepareRequest.parse.parameters.count, 0)
+            XCTAssertEqual(prepareRequest.parse.parameters.first, .text)
             guard case .preparedStatement(let name) = prepareRequest.describe else {
                 fatalError("Describe should contain a prepared statement")
             }
@@ -555,7 +608,7 @@ class PostgresConnectionTests: XCTestCase {
 
             let prepareRequest = try await channel.waitForPrepareRequest()
             XCTAssertEqual(prepareRequest.parse.query, "SELECT datname FROM pg_stat_activity WHERE state = $1")
-            XCTAssertEqual(prepareRequest.parse.parameters.count, 0)
+            XCTAssertEqual(prepareRequest.parse.parameters.first, .text)
             guard case .preparedStatement(let name) = prepareRequest.describe else {
                 fatalError("Describe should contain a prepared statement")
             }
@@ -600,7 +653,8 @@ class PostgresConnectionTests: XCTestCase {
             database: "database"
         )
 
-        async let connectionPromise = PostgresConnection.connect(on: eventLoop, configuration: configuration, id: 1, logger: self.logger)
+        let logger = self.logger
+        async let connectionPromise = PostgresConnection.connect(on: eventLoop, configuration: configuration, id: 1, logger: logger)
         let message = try await channel.waitForOutboundWrite(as: PostgresFrontendMessage.self)
         XCTAssertEqual(message, .startup(.versionThree(parameters: .init(user: "username", database: "database", options: [], replication: .false))))
         try await channel.writeInbound(PostgresBackendMessage.authentication(.ok))
